data-capture/on-car/requirements.txt
python-can>=4.3.1
pynmea2>=1.19.0
pyserial>=3.5
requests>=2.31.0
smbus2>=0.4.2

